requirements.txt
streamlit
openai
faiss-cpu
python-dotenv
PyMuPDF
fastapi
uvicorn
sqlalchemy
psycopg2-binary 